requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.15
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pillow==11.0.0
publicsuffixlist==1.0.2.20241225
pylookyloo==1.26.1
pymisp==2.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzbar==0.1.9
requests==2.32.3
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
bs4
chardet
pyheif==0.8.0
feedparser
gunicorn